// src/Models/Custodian/custodian.ts
export class Custodian{
    SrNo: number;
    Result: number;
    CountryCode: string;
    CountryName: string;
    CustodianCode: string;

    CustodianName: string;
    PMSCode: string;
    PMSName: string;
    PMSAccountNumber: string;
}

export class Country{
    CountryCode: string;
    CountryName: string;
}

export class PMS{
    PMSCode: string;
    PMSName: string;
    PMSAccountNumber: string;
}